Alert component. Interfaces and hard-code styles
