+++ Clients/src/domain/interfaces/iWidget.ts
@@ -1,0 +1,9 @@
+/**
+ * Props for the CloseButton component.
+ *
+ * @interface CloseIconProps
+ * @property {string} text - The color of the close icon.
+ */
+export interface CloseIconProps {
+  text: string;
+}
